+- update to patches.fixes/linux-post-2.6.4-20040314
suse-commit: 8bee4239084478bfda02c96327fb30bce98d7c0a

+++ kernel/sched.c
@@ -915,9 +915,9 @@ unsigned long nr_uninterruptible(void)
 	return sum;
 }
 
-unsigned long nr_context_switches(void)
+unsigned long long nr_context_switches(void)
 {
-	unsigned long i, sum = 0;
+	unsigned long long i, sum = 0;
 
 	for_each_cpu(i)
 		sum += cpu_rq(i)->nr_switches;
